SP: refactor the code of post function
