+++ app/models/board.py
@@ -5,7 +5,7 @@ class Board(db.Model):
     board_id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String)
     owner = db.Column(db.String)
-    card_id = db.Column(db.Integer, db.Foreignkey('board.board_id'))
+    card_id = db.Column(db.Integer, db.ForeignKey('board.board_id'))
 
 
     def to_dict_board(self):
